staging: wilc1000: use 'u16' data type for config id parameter
Cleanup patch to use the correct data type of u16 for keeping the WID
value in 'wilc_cfg_word' & 'wilc_cfg_str' structure.

+++ wilc/wilc_wlan_cfg.h
@@ -27,12 +27,12 @@ struct wilc_cfg_word {
 };
 
 struct wilc_cfg_str {
-	u32 id;
+	u16 id;
 	u8 *str;
 };
 
 struct wilc_cfg_bin {
-	u32 id;
+	u16 id;
 	u8 *bin;
 };
